requirements.txt
# DataMorpher 2.0 - Python Dependencies
# Notes:
# - Torch/TorchVision wheels vary by CUDA/CPU. See README for install tips if pip fails.
# - Some integrations are optional and guarded by runtime checks.

# Core runtime
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0.1

# Async + IO
aiohttp>=3.9.0
aiofiles>=23.2.1

# Data/Science stack
numpy>=1.26.0
pandas>=2.0.0

# Imaging
Pillow>=10.0.0
opencv-python-headless>=4.8.0.76
albumentations>=1.3.1

# ML / DL
torch
torchvision
transformers>=4.41.0
scikit-learn>=1.4.0

# Local LLM (prompt parser)
llama-cpp-python>=0.2.90

# Web UI
streamlit>=1.34.0

# Crawlers / Sources
google-search-results>=2.4.2  # Provides 'from serpapi import GoogleSearch'
praw>=7.7.0                   # Reddit (optional; enabled when credentials present)
yt-dlp>=2024.8.6              # YouTube (optional)
ffmpeg-python>=0.2.0          # Audio extraction for YouTube (optional)

# Optional: Face expression engine (DeepFace)
# Used when selecting "DeepFace" curation in the UI; otherwise not required.
deepface>=0.0.93

# Dev / Tests
pytest>=8.0.0
